# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Role-play API",
        "description": "Документация API для role-play",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Список пользователей",
                "operationId": "ae587d6cf94711fcfeee409b6944e7fe",
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "description": "Фильтр по login",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кол-во на страницу",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "without_paginate",
                        "in": "query",
                        "description": "Без пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/show": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получить пользователя",
                "operationId": "a967e482f7d163676842805e88e0893f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/store": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Создать пользователя",
                "operationId": "8fbd70b63e27ba7e7cc8409b1f06a21b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/update": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Обновить пользователя",
                "operationId": "a9ab9c7f3653b3271048530c1d848be0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/delete": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Удалить пользователя",
                "operationId": "e32c826747e5381fe0f181b43bad3103",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалён"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreUserRequest": {
                "required": [
                    "login",
                    "email",
                    "password",
                    "role"
                ],
                "properties": {
                    "login": {
                        "type": "string",
                        "maxLength": 32,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "example": "john_doe_123"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "john@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "StrongP@ssw0rd"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "StrongP@ssw0rd"
                    },
                    "avatar": {
                        "description": "Изображение (jpeg, png, jpg, gif, svg)",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "role": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "UpdateUserRequest": {
                "properties": {
                    "login": {
                        "type": "string",
                        "maxLength": 32,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "example": "john_doe_123"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "john@example.com"
                    },
                    "avatar": {
                        "description": "Изображение (jpeg, png, jpg, gif, svg)",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "role": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User",
                "description": "Пользователь",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "login": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Bearer authentication using JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}